dev-requirements.txt
-e https://github.com/ckan/ckanext-pages.git
pytest-ckan
